+++ lce/src/main/java/com/laimiux/lce/UC.kt
@@ -8,6 +8,17 @@ interface UC<out C> {
     companion object {
         fun loading(): UC<Nothing> = Type.Loading()
         fun <T> content(value: T): UC<T> = Type.Content(value)
+
+        /**
+         * Returns [UC.loading] if [content] is null, otherwise returns [UC.content].
+         */
+        fun <C : Any> fromNullable(content: C?): UC<C> {
+            return if (content == null) {
+                loading()
+            } else {
+                content(content)
+            }
+        }
     }
 
     fun isLoading(): Boolean