+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
-## [0.1.2] - TBD
+## [0.2.0] - TBD
 - Added LC type.
 - Added `foldLoading` extension method to `LCE`, `UCE` and `UCT` types. 
 - Added `foldError` extension method to `LCE`, `UCE` and `UCT` types.
-
+- Added `fromNullable` factory method to `LCE`, `UCE`, `UCT`, `LC`, `UC` types.
 
 ## [0.1.1] - January 5, 2021
 - Added UCE type.